chore: Some category fixes
