+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/PlaceListProvider.java
@@ -7,6 +7,7 @@ package com.wynntils.services.mapdata.providers.builtin;
 import com.wynntils.services.mapdata.providers.json.JsonMapAttributes;
 import com.wynntils.services.mapdata.providers.json.JsonMapLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
+import com.wynntils.utils.StringUtils;
 import com.wynntils.utils.mc.type.Location;
 import java.util.ArrayList;
 import java.util.List;
@@ -38,6 +39,26 @@ public class PlaceListProvider extends BuiltInProvider {
     public static final class PlaceLocation extends JsonMapLocation {
         private PlaceLocation(String featureId, String categoryId, JsonMapAttributes attributes, Location location) {
             super(featureId, categoryId, attributes, location);
+        }
+
+        public enum PlaceType {
+            PROVINCE("province"),
+            CITY("city"),
+            TOWN_OR_PLACE("town-or-place");
+
+            private final String mapDataId;
+
+            PlaceType(String mapDataId) {
+                this.mapDataId = mapDataId;
+            }
+
+            public String getMapDataId() {
+                return mapDataId;
+            }
+
+            public String getName() {
+                return StringUtils.capitalized(mapDataId);
+            }
         }
     }
 }
